requirements.txt
absl-py>=0.7.1
numpy>=1.18.0
six>=1.12.0
dm-tree>=0.1.1
wrapt>=1.11.1
tabulate>=0.7.5
